docker-mcp-setup/requirements.txt
mcp[sse]
proxmoxer
requests
uvicorn
python-dotenv
